Deleted annotate_c
