+++ audino/frontend/src/containers/modal.js
@@ -16,59 +16,59 @@ import EditProjectForm from './forms/editProjectForm';
 import DownloadDataForm from './forms/downloadDataFrom';
 
 const FormModal = props => {
-  console.log('INSIDE FORMMODAL', props.formType);
+  const {
+    show,
+    onExited,
+    onHide,
+    title,
+    formType,
+    userId,
+    projectId,
+    labelId,
+    labelValueId,
+    api_key,
+    projectName
+  } = props;
   return (
     <Modal
-      show={props.show}
-      onExited={props.onExited}
-      onHide={props.onHide}
+      show={show}
+      onExited={onExited}
+      onHide={onHide}
       size="lg"
       aria-labelledby="contained-modal-title-vcenter"
       centered
     >
       <Modal.Header closeButton>
-        <Modal.Title id="contained-modal-title-vcenter">{props.title}</Modal.Title>
+        <Modal.Title id="contained-modal-title-vcenter">{title}</Modal.Title>
       </Modal.Header>
       <Modal.Body>
-        {props.formType === 'NEW_USER' ? <CreateUserForm authNeeded="true" /> : null}
-        {props.formType === 'NEW_PROJECT' ? <CreateProjectForm /> : null}
-        {props.formType === 'EDIT_USER' ? <EditUserForm userId={props.userId} /> : null}
-        {props.formType === 'Edit_PROJECT' ? <EditProjectForm projectId={props.projectId} /> : null}
-        {props.formType === 'DELETE_USER' ? (
-          <DeleteUserForm userId={props.userId} onDelete={props.onExited} />
+        {formType === 'NEW_USER' ? <CreateUserForm authNeeded="true" /> : null}
+        {formType === 'NEW_PROJECT' ? <CreateProjectForm /> : null}
+        {formType === 'EDIT_USER' ? <EditUserForm userId={userId} /> : null}
+        {formType === 'Edit_PROJECT' ? <EditProjectForm projectId={projectId} /> : null}
+        {formType === 'DELETE_USER' ? <DeleteUserForm userId={userId} onDelete={onExited} /> : null}
+        {formType === 'MANAGE_PROJECT_USERS' ? (
+          <ManageUsersProjectForm projectId={projectId} />
         ) : null}
-        {props.formType === 'MANAGE_PROJECT_USERS' ? (
-          <ManageUsersProjectForm projectId={props.projectId} />
+        {formType === 'NEW_LABEL' ? <CreateLabelForm projectId={projectId} /> : null}
+        {formType === 'EDIT_LABEL' ? (
+          <EditLabelForm projectId={projectId} labelId={labelId} />
         ) : null}
-        {props.formType === 'NEW_LABEL' ? <CreateLabelForm projectId={props.projectId} /> : null}
-        {props.formType === 'EDIT_LABEL' ? (
-          <EditLabelForm projectId={props.projectId} labelId={props.labelId} />
+        {formType === 'NEW_LABEL_VALUE' ? <CreateLabelValueForm labelId={labelId} /> : null}
+        {formType === 'EDIT_LABEL_VALUE' ? (
+          <EditLabelValueForm labelId={labelId} labelValueId={labelValueId} />
         ) : null}
-        {props.formType === 'NEW_LABEL_VALUE' ? (
-          <CreateLabelValueForm labelId={props.labelId} />
+        {formType === 'DELETE_LABEL_VALUE' ? (
+          <DeleteLabelValueForm labelId={labelId} labelValueId={labelValueId} />
         ) : null}
-        {props.formType === 'EDIT_LABEL_VALUE' ? (
-          <EditLabelValueForm labelId={props.labelId} labelValueId={props.labelValueId} />
+        {formType === 'DELETE_LABEL' ? (
+          <DeleteLabelForm labelId={labelId} projectId={projectId} />
         ) : null}
-        {props.formType === 'DELETE_LABEL_VALUE' ? (
-          <DeleteLabelValueForm labelId={props.labelId} labelValueId={props.labelValueId} />
+        {formType === 'UPLOAD_DATA' ? (
+          <UploadDataForm projectId={projectId} projectName={projectName} apiKey={api_key} />
         ) : null}
-        {props.formType === 'DELETE_LABEL' ? (
-          <DeleteLabelForm labelId={props.labelId} projectId={props.projectId} />
-        ) : null}
-        {props.formType === 'UPLOAD_DATA' ? (
-          <UploadDataForm
-            projectId={props.projectId}
-            projectName={props.projectName}
-            apiKey={props.api_key}
-          />
-        ) : null}
-        {props.formType === 'DOWNLOAD_DATA' ? (
-          <DownloadDataForm
-            projectId={props.projectId}
-            projectName={props.projectName}
-            apiKey={props.api_key}
-          />
+        {formType === 'DOWNLOAD_DATA' ? (
+          <DownloadDataForm projectId={projectId} projectName={projectName} apiKey={api_key} />
         ) : null}
       </Modal.Body>
     </Modal>
